Put everything related to facebook api into separate module
